# apps/web/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ASKO API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/repairers": {
      "post": {
        "operationId": "RepairerController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRepairerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepairerResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Repairers"
        ]
      },
      "get": {
        "operationId": "RepairerController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedRepairersResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Repairers"
        ]
      }
    },
    "/repairers/{id}": {
      "patch": {
        "operationId": "RepairerController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRepairerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepairerResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Repairers"
        ]
      },
      "get": {
        "operationId": "RepairerController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepairerResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Repairers"
        ]
      }
    },
    "/repairers/location": {
      "post": {
        "operationId": "RepairerController_updateLocation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLocationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepairerResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Repairers"
        ]
      }
    },
...
    "/repairers/city/{city}": {
      "get": {
        "operationId": "RepairerController_findByCity",
        "parameters": [
          {
            "name": "city",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepairerListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Repairers"
        ]
      }
    },
...
    "/reviews": {
      "post": {
        "operationId": "ReviewController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Reviews"
        ]
      }
    },
...
    "/reviews/{id}/images/{imageId}": {
      "delete": {
        "operationId": "ReviewController_removeImage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Reviews"
        ]
      }
    },
    "/reviews/{id}/images": {
      "get": {
        "operationId": "ReviewController_findImages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Reviews"
        ]
      },
      "post": {
        "operationId": "ReviewUploadController_uploadReviewImage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Review uploads"
        ]
      }
    },
...
    "/reviews/rating/repairer/{repairerId}": {
      "get": {
        "operationId": "ReviewController_findRepairerRating",
        "parameters": [
          {
            "name": "repairerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatingResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Reviews"
        ]
      }
    },
    "/reviews/repairer/{repairerId}": {
      "get": {
        "operationId": "ReviewController_findByRepairer",
        "parameters": [
          {
            "name": "repairerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedReviewsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Reviews"
        ]
      }
    },
    "/reviews/request/{requestId}": {
      "get": {
        "operationId": "ReviewController_findByRequest",
        "parameters": [
          {
            "name": "requestId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Reviews"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthProvider": {
        "type": "string",
        "enum": [
          "EMAIL",
          "PHONE",
          "GOOGLE",
          "VK",
          "YANDEX"
        ]
      },
      "Role": {
        "type": "string",
        "enum": [
          "super_admin",
          "admin",
          "dealer",
          "manager",
          "repairer",
          "user"
        ]
      },
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "middleName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "providers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthProvider"
            }
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "providers",
          "roles",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "EmptyResponseDto": {
        "type": "object",
        "properties": {}
      },
...
      "CloudinaryImageDto": {
        "type": "object",
        "properties": {
          "public_id": {
            "type": "string"
          },
          "version": {
            "type": "number"
          },
          "signature": {
            "type": "string"
          },
          "width": {
            "type": "number"
          },
          "height": {
            "type": "number"
          },
          "format": {
            "type": "string"
          },
          "resource_type": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "secure_url": {
            "type": "string"
          },
          "original_filename": {
            "type": "string"
          }
        },
        "required": [
          "public_id",
          "version",
          "signature",
          "width",
          "height",
          "format",
          "resource_type",
          "url",
          "secure_url",
          "original_filename"
        ]
      },
      "ImageObjDto": {
        "type": "object",
        "properties": {
          "original": {
            "$ref": "#/components/schemas/CloudinaryImageDto"
          },
          "thumbnail": {
            "$ref": "#/components/schemas/CloudinaryImageDto"
          },
          "medium": {
            "$ref": "#/components/schemas/CloudinaryImageDto"
          },
          "large": {
            "$ref": "#/components/schemas/CloudinaryImageDto"
          }
        },
        "required": [
          "original"
        ]
      },
      "ImageRecordDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "imageJson": {
            "$ref": "#/components/schemas/ImageObjDto"
          },
          "alt": {
            "type": "string"
          },
          "order": {
            "type": "number"
          },
          "ownerType": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "imageJson",
          "order"
        ]
      },
      "ImageListResponseDto": {
        "type": "object",
        "properties": {
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImageRecordDto"
            }
          }
        },
        "required": [
          "images"
        ]
      },
...
      "ImageResponseDto": {
        "type": "object",
        "properties": {
          "image": {
            "$ref": "#/components/schemas/ImageRecordDto"
          }
        },
        "required": [
          "image"
        ]
      },
...
      "RepairRequestStatus": {
        "type": "string",
        "enum": [
          "pending",
          "paid",
          "assigned",
          "accepted",
          "en_route",
          "in_progress",
          "awaiting_completion",
          "completed",
          "refused",
          "cancelled",
          "paused",
          "refund_requested",
          "refunded"
        ]
      },
...
      "RepairerRecordDto": {
        "type": "object",
        "properties": {
          "currentRequestStatus": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RepairRequestStatus"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "specializations": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "city": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "rating": {
            "type": "number"
          },
          "completedRepairs": {
            "type": "number"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "lastLocationUpdate": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          },
          "activeRequestCount": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "specializations",
          "city",
          "isActive",
          "completedRepairs",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateRepairerDto": {
        "type": "object",
        "properties": {}
      },
      "RepairerResponseDto": {
        "type": "object",
        "properties": {
          "repairer": {
            "$ref": "#/components/schemas/RepairerRecordDto"
          }
        },
        "required": [
          "repairer"
        ]
      },
      "UpdateRepairerDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateLocationDto": {
        "type": "object",
        "properties": {}
      },
      "PaginatedRepairersResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RepairerRecordDto"
            }
          },
          "overallCount": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "data",
          "overallCount",
          "page",
          "limit"
        ]
      },
      "RepairerListResponseDto": {
        "type": "object",
        "properties": {
          "repairers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RepairerRecordDto"
            }
          }
        },
        "required": [
          "repairers"
        ]
      },
...
      "CreateReviewDto": {
        "type": "object",
        "properties": {}
      },
      "ReviewRecordDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "repairRequestId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "repairerId": {
            "type": "string"
          },
          "rating": {
            "type": "number"
          },
          "comment": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "repairRequestId",
          "userId",
          "repairerId",
          "rating",
          "createdAt"
        ]
      },
      "ReviewResponseDto": {
        "type": "object",
        "properties": {
          "review": {
            "$ref": "#/components/schemas/ReviewRecordDto"
          }
        },
        "required": [
          "review"
        ]
      },
...
      "RatingResponseDto": {
        "type": "object",
        "properties": {
          "average": {
            "type": "number"
          },
          "count": {
            "type": "number"
          }
        },
        "required": [
          "average",
          "count"
        ]
      },
      "PaginatedReviewsResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewRecordDto"
            }
          },
          "overallCount": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "data",
          "overallCount",
          "page",
          "limit"
        ]
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    }
  }
}